+++ app/models/concerns/searchable.rb
@@ -6,30 +6,21 @@ module Searchable
 
   included do
     # Search Scope
-    def self.search(arg, match_start: true)
-      term = \
-        if match_start
-          arg.to_s.downcase.gsub(/ |$/, "%")
-        else
-          arg.to_s.downcase.gsub(/^| |$/, "%")
-        end
-      terms = [term] * search_queries.count
-      where search_queries.join(" or "), *terms
-    end
-
     def self.search_any_order(args)
       terms = args.to_s.split(/\s/).collect do |arg|
         arg.to_s.downcase.gsub(/^| |$/, "%")
       end
-      where(
-        (search_queries.collect { |s| [s] * terms.count }).flatten.join(" or "),
-        *(terms * search_queries.count)
-      )
+      queries = [concat_ws] * terms.count
+      where queries.join(" AND "), *terms
     end
 
-    def self.search_queries
-      searchable_attributes.collect do |searchable_attribute|
-        "#{table_name}.#{searchable_attribute} ILIKE ?"
+    def self.concat_ws
+      "concat_ws(' ', #{full_attributes.join(", ")}) ILIKE ?"
+    end
+
+    def self.full_attributes
+      searchable_attributes.collect do |attribute|
+        "#{table_name}.#{attribute}"
       end
     end